BUG#33378364: group_replication.gr_perfschema_group_communication_information
When a member joins while a group action[1] is ongoing, since its
configuration may conflict with the ongoing coordinated change, the
member will be removed from the group. That occurrence is logged on
all members.

On the joining member the logged message is:
  [ERROR] [MY-013207] [Repl] Plugin group_replication reported: 'A
  member cannot join the group while a group configuration operation
  is running.'
  [System] [MY-011504] [Repl] Plugin group_replication reported:
  'Group membership changed: This member has left the group.'
Also the following error message is thrown by START
GROUP_REPLICATION:
  The server is not configured properly to be an active member of
  the group. Please see more details on error log.

On the other members the logged message is:
  [Warning] [MY-013208] [Repl] Plugin group_replication reported: 'A
  member is joining the group while a group configuration operation
  is occurring. The member will now leave the group'

Though, the message on the others members was being logged in
situations on which there were no joining members. Like when `SELECT
group_replication_set_communication_protocol()` is executed,
that does trigger a view delivery on which there are no joining
members.

To solve the above issue, now the message on the other members is
logged when there are indeed joining members.

[1] https://dev.mysql.com/doc/refman/8.0/en/group-replication-configuring-online-group.html

RB: 27412