Bug#33611740: Hypergraph optimizer loses conditions in pushdown
When the hypergraph optimizer attempts to push join conditions as far
down as possible, it may leave multi-equalities too high in the join
tree and eventually lose some of the simple equalities represented by
the multi-equality.

Fixed by making PushDownAsMuchAsPossible() always try to push as much
as possible of the multi-equality further down the tree.

Change-Id: If7e62dd7abc37d7dc597b9478100381765344923