WL#14370, Prepare for semi-join nest to be 'nested' as well.
Currently only the (outer) join-nest could be nested, that is having an
OUTER-nest as a sub-nest inside another OUTER join nest.

However, the AccessPath structures might represent nested semi-joins
as well, which this patch prepare the nest-building code in
::is_pushable_as_root() for:

The AQP method get_first_sj_upper() is introduced, intended to return
the sj_nest-upper which this sj_nest is nested inside. As the AQP
is *still* based on the QEP_TAB's, not Accesspath, its implementation
currently return '-1' -> not nested.

The nest forming code in is_pushable_with_root() is extended to handle
multiple semi_joins nest.

Note that a slight change in how the sj_nest bitmap is set up is
introduced as well. Instead if holding all all semi_joined tables from
first_sj_inner upto 'this' table, it is extended to contain all
tables between first_sj_inner .. last_sj_inner (which always include
'this' table).

As get_first_sj_upper() will always still return '-1' -> 'not nested',
there should be no effective changes as a result of this patch.

Change-Id: I16c9ee73bce34bf3d8922b76def8b2734ad00b62