Bug#33300235 cs fails to rejoin the old primary cluster back to cs after forced failover
If a few conditions are met:
-target_cluster is a Replica Cluster
-target_cluster and former Primary Cluster both get invalidated in the metadata
-the invalidated_cluster_policy is drop_all
-the Router still can access former Primary Cluster

the Router would keep using former Primary Cluster for periodic metadata updates
due to the bug in the code (returning too soon from the update function before
saving the new highest view_id)

This patch fixes the bug (now if the target_cluster got invalidated the new
view_id is honored as expected).
It also splits the update function into smaller chunks to make it easier
to follow the logic.

RB: 27416