Bug#33506241  sql produces different resultsets when different internel storage engines r used
Contributed by Brian Yue

SELECT DISTINCT on a VARCHAR returns 2 distinct values when the
coloumn contains one empty string and one string which consists of a
single whitespace character. This occurs only when the TempTable
storage engine is used. This is incorrect because the VARCHAR field
uses a utf8_bin collation with PAD SPACE, so trailing spaces should
be ignored while selecting distinct values.

For temporary tables created using the TempTable storage engine, the
two strings are hashed to different values. The TempTable hash
function always hashes zero-length data to 0. So the empty string is
hashed to 0 and the non-empty string's whitespace character is used
to generate a nonzero hash value.

This is fixed by removing the special handling for zero-length data
in the TempTable hash function.

Change-Id: Ibea34bb4176d923c3acdba2e66350e37b7936903