WL#14370, prepare join_pushdown MTR test for AccessPath usage.
There will be some changes to what can be pushed and not as
a result of changing to use the AccessPath structure as basis for
the AQP.

Some of these changes are better understood when using the 'tree'
explain format, which are the main part of this patch.

Also added some WL#14370 specific comments to the test result -
mainly to explain changes to happen as result of next patches.

Change-Id: Ie311f96c99548b31e5b27a65921bbe98b9ae2d6e