wl#14370, Remove obsolete code for handling FOUND_MATCH-triggers.
When AccessPaths are generated from QEP_TABs all trigger conditions
are eliminated, and the condition parts are converted into FILTERs put
on top of the AccessPath branch which covers the triggering tables.
Such FILTERS will be checked with the has_condition_inbetween()
method added in previous patch.

Thus, the code inside ha_ndbcluster_push.cc which walked the
condition expression, trying to identify such triggers, are now
obsolete. (As there are no such triggers to be found anymore)

There is also a special antijoin_null_cond variant of the
FOUND_MATCH trigger, used to identify the particular join
condition as an ANTI_JOIN. That is now explicitely set in
the AccessPath by setting JoinType = ANTI_JOIN. Thus that
became obsolete as well.

Patch remove this obsolete code.

Change-Id: I7acc2fcdb4042acd5cd6d325e848f8d0f3765fbb