WL#14370, remove obsolete code for setting 'spill_to_disk = false'
The 'spill_to_disk' variable decide if a HASH_JOIN is allowed to
write chunks of the hash-buckets to disk if all of it can't be
kept in memory (too large)

For pushed-joins we can only allow a 'spill_to_disk' if the entire
pushed join is within the AccessPath branch which goes into the
hash-bucket. (See reasoning about this in comments being deleted,
now partially moved to acceptPsuhedJoins and family)

The original spill_to_disk logic was on the call paths when
we build an AccessPath from the QEP_TAB. Prior to this WL we had
already generated the pushed joins (from QEP_TABs) at this point.

With this WL pushed joins are generated _after_ the AccessPaths has
been set up, so the original code for spill_to_disk is effectively
defuncted. Instead this is now already taken care of in
fixupPushedAccessPaths() when it find an AccessPath::HASH_JOIN.

Thus the original code for this can be removed.

Also note that in the original implementation we could
'RejectHashJoin', that is; force them into a nested-loop-join
instead, even if the optimizer had decided for a hash-join.
As we now operate on a 'finalized' AccessPath we can't do
such major queryplan changes any more. This explains a few
of the query plan changes seen in the previous huge patch which
introduced the Accesspath based AQP. (Where explain now
shows a hash-join instead of a nested-loop join)

Change-Id: I4c15457b5c701921f940323980dd38ae3ff55ac3