Bug#33305617: Assertion `table_independent_conds == nullptr' failed.
If an aggregation function is evaluated during optimization,
it is presumed that where condition is already evaluated.
Optimizer could therefore remove the where condition.
Before the removal, it verifies if the where condition
has any table independent conditions by making a call
to make_cond_for_table(). However, if a condition is
considered expensive i.e has stored procedures, this
will incorrectly conclude that we indeed have a table
independent condition leading to an assert in debug builds.
We fix it by excluding expensive conditions in the
call to make_cond_for_table().

We also see wrong results when we have a constant
condition but is expensive to evaluate like f() = 1
where function f() does not use any tables.
For most cases, if a condition is constant for execution
it gets evaluated when optimizing the where condition.
Evaluation is skipped if it's considered expensive.
But we are also skipping it while optimizing an
implicitly grouped query leading to wrong results.
We now fix it by evaluating such conditions before
aggregating.

Change-Id: Id8702f134c8dcbd845376f63cb20b4bc6ac0e963