Bug#33616957: Parser error for TIME values in binary protocol
The synopsis here is slightly misleading. Bug#33539844 unintentionally
fixed a problem with insertion of datetime values containing a time
zone offset into a TIME column using prepared statements. Before this
bug fix, the time zone offset was not processed.
(This is not a problem for regular TIME literals, where a time zone
offset cannot be specified).

However, this left a difference with execution of an equivalent regular
statement, where a time zone offset would still be ignored.

The fix is to also recognize and adjust for time zone offsets
whenever a timestamp value with a time zone offset is inserted into
a TIME column.

The test case also adds insertion from a datetime column, to ensure
that a datetime value from a column is treated identically to a
literal value or a string dynamic parameter.

Change-Id: Ifd277d6c861fb2f615c5d77270f39717d3a47045