Bug#33615893: EXPLAIN statement crashes in get_fields_used()
For a query, a range scan can be picked first and optimizer could
decide that the same range scan can also be used to skip ordering.
In some cases, when the requested order is not the same as index order,
it needs to do a backward index scan. If the requested ordering is
on the "keypart" that is not already used by the range scan, we
update the number of used key parts for the range scan to reflect the
change. But the key part information is not updated.
This leads to failure later while accessing key part information
based on the number of key parts used.

For a reverse scan, the information about the number of key parts
used to provide order is needed only to know if we are using
the extended key parts in a secondary index (In case of innodb,
secondary indexes are always extended with primary key). Using
extended keyparts would need different flags set while retrieving
rows in case of reverse range scan. So the fix is to make a note
of it when range scan is marked for backward index scan. While
creating/reading from the iterator, use it to set the correct flags.

Change-Id: I972aab70fceaffe0029d18b705350f45fa5cf604