WL#14370, use 'query_scope' bitmap to calculate set of 'const_expr_tables'.
Step to remove usage if AQP methods still being dependant on QEP_TABs.
When preparing a pushed condition we supply a bitmap argument containing
those preceding tables which cen be referred as constants at the time of
execution.

Prior to this WL that was calculated as all preceeding tables, except
if a 'join buffer' was referred, which used the QEP_TAB dependant method
uses_join_cache().

With this WL the AQP directly provide a bitmap set of tables
being within our 'scope'. See comments in abstract_query_plan
implementation regarding the scoping definitions.

As a side effect we may also push a few more conditions.

Change-Id: I2edd521b8969a27e999c6e444dcb7c56b1322f50