WL#14370, replace AQP::is_sj_firstmatch() with ::is_semi_joined().
is_sj_firstmatch() is one of the still remaining AQP methods depending
on QEP_TAB's - we need to remove these.

is_sj_firstmatch() check the QEP_TAB sj_strategy to be 'firstMatch',
which is only one of many semi-join strategies exposed by QEP_TAB's.

In the AccessPath representation of the query plan, there is no such
thing as a semi-join strategy at all - so there is no direct mapping
of this into the AccessPath based AQP.

Instead we need to use the Join_nest / Join_scope structure to check
if a pushed child candidate is within a SEMI_JOIN nest, _and_ refer
a parent outside of this nest, which is implemented by is_semi_joined().

As the original is_sj_firstmatch() only allowed for a firstMatch
semi-join to be pushed, we will now open up for other semi-join
variants to be pushed as well - see changed MTR results.

However, this also causes a couple of the tests to return incorrect
results. This will need fixes coming in the next patches.

Change-Id: I821f07a1aca99db111e38d7ceceee38c339c3a5f