WL#14370, fix, don't allow field substitute into semi-join nests.
Fix one of the two queries which started to fail after previous patch:

is_field_item_pushable() check whether a child's Field reference,
used to generate an index key, refers a table within the allowed
'query_scope'. In addition it also use the Item_equal set to find
Fields from other tables which can be substitutes for the referred field.
(Item_equal holds a list of Fields know to be equal.)

It turns out that we are not strict enough in checking which of the
Item_equal fields we allow to be 'substitutes'.

In addition to check that substitutes are within our join_scope,
we also need to make sure that we do not substitute with a fields
which are within a semi-join: That would effectively make this child
table a semi-join'ed table as well!

Change-Id: I324748dd9989dfa11bac857bf1eeaa3efc9e397e