WL#14370, Enable join-pushdown for Hypergraph optimizer.
The HG optimizer generate the AccessPaths directly, without
going via the QEP_TAB structure. Thus, it never reach the
existing push_to_engine() entry point.

Patch add a call to push_to_engine() in the HG call path.

NOTE: We had to disable Item_equal usage for HG optimizer query
plans. Could possibly try to enable that later in a bug fix, but
for the moment the HG optimizer did not provide a usable Item_equal.

NOTE: filesort had an assert that no tables in the filesort list
could be pushed. That assert turned out to be too strict. Also note
that originally that assert was less strict, and checked only a single table.
It seems like some of the Iterator rework of the code made this assert
stricter than required.

Assert was replaced with a check for any child tables requiring a
specific 'use_order()' in is_pushable_as_child().

Change-Id: I9b2d10fc118c0bd59ebadf8ed08e3b56c2b94a80