BUG#32327411 : OFFLINE KEYRING MIGRATION FROM KEYRING_FILE REPORTS SUCCESS, BUT MIGRATES NO KEYS
Problem: For empty or non-existing keyring sources, the offline keyring
migration reports success, but no appropriate warning message is shown.
If the user enters a relative path for the --keyring-file-data variable,
the source keyring is looked in the basedir, as offline keyring
migration
doesn't support relative path, entering a relative path might result in
empty keyring source generation, and transferring keys from it.

Fix: Now the migration process will check for any error during the key
fetch
, if there's no error, and there are no keys, then a warning message is
shown that keyring has no keys in it.
The test will try to migrate the keys from an empty keyring source, and
looks for the warning message in the server log.

RB#27401