Bug#33616019: Hypergraph: Assertion `static_cast<size_t>(bit_num) < x.capacity()' failed
When a zero-rows access path for an impossible range is created, its
filtered_predicates bitset is cleared by copying an empty
OverflowBitset into it. The empty OverflowBitset is inlined and has
capacity 63. This results in assert failures later when trying to
check if any bits higher than bit 63 is set. This can happen if there
are more than 63 predicates in the query.

Fixed by creating an empty OverflowBitset with large enough capacity
for the actual number of predicates in the query.

Change-Id: Id1e3c0c4ddcc7493d5b4a0f6ef9e433231d871bd