Bug#33653934: Assert in main.index_merge_myisam - 32-bit, --hypergraph
Range scans that were children of an index merge did not initialize
the "reverse" bit of their access paths, so it was not deterministic
if they ended up as forward scans or reverse scans. IndexMergeIterator
expects that all its children are IndexRangeScanIterators, and it
failed fail when it tried to use a ReverseIndexRangeScanIterator as an
IndexRangeScanIterator.

Fixed by initializing the "reverse" bit to false when creating the
index range scan access paths.

Change-Id: I9c790d07d990a7e5c4215529baf73338c2943829