Bug#33655246 - Cleanup shared quarantine on shutdown issue
Issue
=====
Each async quarantine handler must stop when the Router is shutting down. To
determine if the router is shutting down currently we are checking for the
state of a routing plugin, but it is possible that the shared quarantine
could outlive some routing plugins - that leads to accessing a destructed
object and undefined behavior.

Fix
===
Do not check if the routing plugin is running but instead check if the shared
quarantine is shutting down.

RB: 27464
Reviewed by: Andrzej Religa <andrzej.religa@oracle.com>