WL#14680 - Removing #defines from InnoDB
Changing #define to constexpr or inline functions.
To have more elegant and modern code and made debugging easier we want to remove #defines and use constexpr or inline functions instead.
In places when it's very hard, or even impossible without major changes in code we leave code as it is. Also we leave old code in places when it doesn't seem so relevant to do the changes.

RB#26491