Bug #33664929 Reduce size of RelWithDebInfo binaries
Packages for some platforms are very large, and this is due to large
binary sizes.  This in turn is caused by large amount of debug
information, which often is of little interest to the user.  On
affected platforms the mysqld binary may be one Gb or more, while on
other platforms it's "only" about 70 Mb.

This patch introduces a new cmake option, visible only for gcc/clang builds on Unix:
  MINIMAL_RELWITHDEBINFO
which will append " -g1" to CMAKE_CXX_FLAGS_RELWITHDEBINFO
It is default ON for STANDALONE and TAR builds that do not specify any
sanitizer or valgrind flags, OFF for other build types.

See also
Bug #33663811 Ship standard debuginfo package for SLES|openSUSE RPM packages
which splits MySQL packages and introduces debuginfo packages for RPM builds.

With "-g1" rather than "-g" the total size of a build directory is
reduced from roughly 20G to 5.5G for a gcc RelWithDebInfo build.
Total build time is reduced by about 25%.

Change-Id: I8a507b0137445c5c476ab60047a5a9ae9c24c0c3