Bug #33045204 EQUI-HEIGHT HISTOGRAM ALWAYS ESTIMATES NON-INTEGRAL EQUALITY PREDICATE TO 1/ROWS [patch 1/2, selectivity fix]
Equi-height histograms consist of a collection of non-overlapping
buckets ([a1, b1], [a2, b2], ...). In addition to storing the endpoints
[a, b], each bucket also contains an estimate of the number of distinct
elements in the bucket together with the total frequency of the bucket.
When estimating the selectivity of an equality predicate for a value
that falls within a bucket [a, b] where a < b, there are (at least) two
different strategies:

1) Assume that the predicate value is uniformly distributed across
   [a, b] and return the average frequency of an element in the bucket
   multiplied by the probability that a random value is equal to an
   element in the bucket.

2) Assume that the predicate value follows the distribution of the data
   and return the average frequency of an element in the bucket.

The problem with the current strategy 1) is that it leads to very low
selectivity estimates when there are many possible values in the
interval [a, b], for instance if a and b are floating point numbers.

This patch implements strategy 2).

Let S(op) denote the selectivity estimate for the operator op,
e.g. S(<) for the < operator. The histogram interface supports S(<),
S(=), and S(>). Several places in the code selectivities were computed
using equalities such as S(<=) = S(<) + S(=) or S(<) = S(<=) - S(=).
Besides being inefficient, with the new semantics of S(=) it also
became somewhat inconsistent. The new patch uses base implementations
of S(<), S(=), and S(>), and computes e.g. S(<=) as 1 - S(>). The idea
is that we use the underlying operators S(<), S(>) when estimating the
selectivity for a range of values, and S(=) when performing
equality-type estimation (=, IN, NOT IN).

A new test is added to verify that the error in selectivity
estimates can be bounded by the number of buckets in the histogram.

Change-Id: I0cafa68c0aa33b0f5288400a0e42120d3d4b6822