Bug#33635637: Regression in FIND_IN_SET() in 8.0.22
This bug was reported as a regression in how variables were used to
look up enumeration values in SET variables using the FIND_IN_SET()
function. However, the problem is actually with how non-const values
were used with FIND_IN_SET(), thus when variables were no longer treated
as const values, it manifest as a regression from WL#9384.

The problem with non-const values is that they are matched directly
against the set in the set variable expression, then returning the
enumeration value from that set. But with const values, the value is
first looked up in the set, then if it is found there, it is further
looked up in the underlying set of the column value, and the
enumeration value from this set is looked up.

Example: We have a column defined as a set 'a', 'b', 'c', 'd', 'e', 'f'
and a column value ['a', 'c', 'f']. When matching a value 'c', we get
the enumeration value 2 when looking at the column value and the value
3 when looking at the column definition. The former value is returned
for a non-const value, the latter for a const value.

When fixing the problem, we first decide that the correct result
is the enumeration value returned when looking at the column definition,
and thus we adjust the result for non-const values.

More details about the solution:

- Member enum_value is renamed as m_enum_value.

- Member enum_bit is removed since it can be recalculated at runtime.

- The evaluation for const values against fields of type SET is not
  changed, except that the enum_bit value must be calculated at runtime.

- The evaluation for other values is split into one part for fields
  of type SET, and one part for other types of expressions.

- For fields of type SET, it is determined if the searched value can
  be found in the underlying set, then if it can be found in the
  expression, and if it the value is presebt in both sets, the
  enumeration from the underlying set is returned.

- For other expressions, the searched value is matched directly against
  the expression, treating the expression as a comma-separated set, and
  the returned value is the enumeration value from this set.

Change-Id: I8bf8c6570edd4e5906d58bed00b531a87549ef21