Bug#33459671 GCC 11 build failure: -Werror=array-bounds
Problem:
--------

The build failed due to a array-bounds error:

error: array subscript 0 is outside array bounds of 'struct
ndb_mgm_cluster_state2[100254043878856258]' [-Werror=array-bounds]
 1566 |   i = -1; ptr--;

Manual code inspection showed that this wasn't really a problem since
ptr was guaranteed to be incremented before its usage in the first
iteration of the loop in question. However, there was scope for
improvement in making things simpler and more obvious to both the
compiler and devs.

Fix:
----

- Avoid decrementing ptr and use 2 pointers instead.
- Fix other weaknesses in the loop code related to failure
  handling.
- Remove workaround that downgraded the warning previously.

Change-Id: I71e283b66ecffc2132d6a8000a7f619532fd6b58