WL#15000 Add ndbcluster plugin threads to P_S
The ndbcluster plugin threads are not visible in
performance_schema.threads table. This makes it hard to find the correct
thread_id for doing other queries and relate it to the operating systems
thread ids.

Add the three threads created inside ndbcluster plugin to
performance_schema.threads table.
The instrumented threads are because of this also named in `ps` and
`top` on platforms that support naming threads.

Change-Id: Id6d7bf8f9f42201d775341fbf8f0a3915605632d