Bug#33664787: hypergraph: Incorrect cost calculation for implicitly grouped query with LIMIT
The estimated number of output rows from an AGGREGATE access path in
an implicitly grouped query is set to the same as the number of input
rows. Since an implicitly grouped query always returns one row, this
estimate may be far off, and make the hypergraph optimizer choose
suboptimal plans. Particularly, in combination with LIMIT, it has been
seen to prefer subplans with low init cost and high total cost, on the
incorrect assumption that the AGGREGATE node will not need to read all
input rows.

This patch sets the estimated number of output rows to 1.0 for
implicitly grouped AGGREGATE paths.

Change-Id: I8b3739f585ac625853fc270df38634851efabf5a