Bug#33619350: Hypergraph: Assertion `it != m_access_paths.end()' failed.
A semi-join failed because the hypergraph optimizer found no plan.

The problem was in code that skipped joins where a join predicate had
been made redundant by sargable join predicates in sub-joins of the
larger join. (A better approach might have been to remove the
redundant join predicate instead of skipping the join, which is also
suggested in a TODO comment in ProposeHashJoin(). But that's
orthogonal to this particular issue.)

The code that analyzes which sargable conditions a join predicate
could be redundant against, has a check to find out whether two
Item_func_eq objects come from the same multiple equality. However,
this check returns true if both objects are single equalities.

Fixed by changing it to check that the Item_func_eq objects actually
come from a multiple equality and, if they do, that they come from the
same multiple equality.

Change-Id: I07880abb3f3b649697fe1cddf58370c3f118c2f2