Bug#33405696 btr_insert_into_right_sibling() leads to ACID violation
The btr_insert_into_right_sibling() function is used in situation where
a record "should" be inserted to the "left" page according to the
record's key relation to the keys stored in non-leaf pages along the
path from the root, yet the "left" page has no room to accomodate it,
and we suspect that the next page ("right") might accomodate it, so
instead of expensive split operation of "left" page, we simply prepend
the inserted record to the "right" page.

This is a rather unusual operation in that it requires updating the
split point between "left" and "right" pages - usual inserts, even at
boundary of the page do not have to adjust anything, because the record
is known to be inserted to the page reached by descending from root and
comparing its key to those in intermediate nodes, so it must be on the
correct side of the split-point.
Similarly, we ususally don't adjust the split point when deleting a row.

But, in our case, we know that the search pointed to the "left" page,
yet we want to put the record on the "right" page, so clearly we need to
adjust the split-point, by moving it to the left. This is achieved by
removing the old pointer leading to the "right" leaf page from its
parent and then adding a new one, with the just inserted key. (This is
done recursively, to properly handle cases where "left" and "right" are
distant "cousins").

The subtle bug was that this also requires adjustment in locks
maintained in Lock System, which protected the gap between last record
of "left" page, and first record of "right" page.
Or rather *gaps*, because Lock System represents it by two distinct
gaps:
- the gap before supremum pseudo-record in "left",
- the gap before first real record in "right".
The two ranges conceptually represent the fragments of key-space each
page is responsible for, given how the navigation algorithm decides to
which of the two to descend when searching for a given key. In other
words, the split-point between the two pages which is used to decide
where to insert a given key, is the also the split-point between the two
gaps.

The code erroneusly called lock_update_split_left(right, next), which
copies locks from right's block gap, to left's block gap.
But what's really needed here is to copy them in the opposite direction!
To see this, observe that moving the split-point to the left, means that
a fragment of the gap which was previously protected by the locks stored
on the left gap, must now be protected by locks on the right page.

Copying the locks from left to right fixes the problem. First the locks
from the gap before left's supremum are copied to the gap before (old)
first record of the right page, and then the record is inserted into
said gap spliting the gap into two, which triggers copying the locks
from the gap before old record (which is now the second) to the gap
before inserted record (which is now the first). This way the locks
meant to be held on the (left) gap are properly replicated to both gaps
- before and after the newly inserted record.

Change-Id: I8e7366e11cb516310480610155a62a7ea41d2cc5
Change-Id: I189546a8da0c352e610abbf8a02ef778136f2975
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>