Bug #33650498: Mixing PS and non-PS syntax breaks --fido-register-factor
Problem: Alter user finish registration sql has mix of PS(prepared statement)
and non PS syntax.

Analysis: Initial design was to implement challenge response as a dynamic
parameter. This was later dropped as part of bug#33008854 stating dynamic
parameters are not allowed for DMLs as it is not SQL standard. However code
cleanup was not done.

Fix: Fixed unwanted code related to PS syntax.

RB#27459