Bug #32019119 SCHEMA DIST SETUP WARNINGS
post-push fix: There is still a possibility that mysqld can loss
connection to NDB after the binlog setup is started and print the
fatal error messages causing the MTR test cases to fail. Since, the
binlog thread can re-try the setup process, any fatal error message
during the setup can be avoided. So, degraded the log level from
error to info for binlog setup failure from the util table creator.

The other log is from the function ndbcluster_binlog_setup_table().
This function is called from various threads and the failure when called
from any thread other than the binlog thread is serious and need to be
reported as an error. So, an optional parameter is introduced, which,
when set, will avoid printing the error message and just returns error
code to the caller and thus leaves the responsibility for logging to
the caller function.

Change-Id: Ic6474e81c8e252962d80bbec1b4d813dde0eae0e