Bug #33675809 	Bump required Doxygen version to 1.9.2, remove all workarounds for earlier ver
The Doxygen was upgraded to 1.9.2 and code was prepared to run with this version in `Bug #33603036 	InnoDB: Fix doxygen errors, enable doxygen for .ic and .i files`.
It and the `WL#14680 - Removing #defines from InnoDB` introduced some workarounds to make it work on 1.9.1 version, which will not be needed anymore after the upgrade to 1.9.2.
Many doxygen errors are fixed, especially around C++17 attributes and usage of `__VA_ARGS__` in macros in argument lists.

Change-Id: I9eb4310bc127444b177648ad119b3e6723cf1a22