Bug#31946448:OPTIMIZE SUBQUERY CAUSES A CORE DUMP
While removing a condition which is always true, a subquery gets removed
even though it is referred somewhere else resulting in server exit.
The subquery that gets removed is a view reference (a field from a merged
derived table). An earlier change (Bug#31216115) had ensured that a
subquery which is a view reference is not removed. However, it missed to
check for view references within a subquery. The fix for the current
problem is to ensure we check all fields in a subquery for view references
thereby ensuring we do not remove them.

Change-Id: If6cb9a307b2195770e16cc294f00983836040268