Bug#33717719 test-fail if port 3306 is in use
Issue
=====

routertest_component_mock_server fails if port 3306 is already in use.

The "xproto_mysqlsh_select_connection_id" test only specifies
the "--xproto" port when starting the mock-server as it only uses the
xproto port.

The mock-server always builds the classic protocol port. If not
specified, it uses 3306.

Change
======

- set --port for the "xproto_mysqlsh_select_connection_id" test

RB: 27514