Bug#33725507: A bug in sql_lex.cc cause mysql-server to fail
This problem needs a few preconditions to hit:

- A derived table is merged into an outer query block (call them
  qb#4 and qb#3).

- An IN subquery predicate is converted to a semi-join (the subquery is
  qb#2 and the outer query block is qb#1).

- The query block qb#4 has an ORDER BY clause with a subquery with an
  outer reference into qb4.

During the semi-join transformation, the underlying query blocks are
inspected and Item::fix_after_pullout() is called for each contained
item. One of these items is a converted subquery predicate (which has
not been transformed to a semi-join), one of these is the above
mentioned outer reference, and when attempting to update dependency
information, the end criterion for the loop is never met, thus we
eventually try to access a query block above the outermost one.

The reason for this is that when merging the derived table, the ORDER BY
clause is moved from qb4 to qb3. However, Item::fix_after_pullout()
is never called for this clause, thus depended_from information is wrong.

The fix for the problem is then to call Item::fix_after_pullout() for
the items in the ORDER BY clause.

Change-Id: I4636171f817fc335edfcf2c4741d3fa594ad4d69