BUG #33302021 Equi-height histogram is not equi-height
MySQL supports the use of equi-height histograms to improve selectivity
estimates. Each bucket in an equi-height histogram for a column should
contain roughly the same number of values (rows). Keeping the buckets
small helps in minimizing the selectivity estimation error.

The current equi-height construction algorithm sometimes places too many
values into the same bucket, potentially leading to large errors in
selectivity estimation.

This patch introduces a new equi-height construction algorithm that
guarantees low error. In addition, the new algorithm adapts to the
distribution of the data to make efficient use of its buckets.

This patch also implements a new estimator for the number of distinct
values in histogram buckets. The new Guaranteed Error Estimator (GEE)
is borrowed from the database literature. It is simple to compute and
has better worst-case error guarantees compared to the old estimator.

Change-Id: Iac16a2754685de888882773972c13192c7a5ff81