Bug#33738447: Assertion `!(used_tabs & (~ read_tables & ~filter_for_table))' failed.
If the hypergraph optimizer adds a materialization step before
aggregation in a prepared statement, it might hit an assert failure in
the second execution of the statement, when estimating the selectivity
of the HAVING clause.

The reason is that the first execution rewrites the HAVING clause to
point into the temporary table during FinalizePlanForQueryBlock(), and
updates its used_tables() cache to reflect the change (see the call to
item->update_used_tables() near the end of
ReplaceMaterializedItems()). At the end of the first execution, the
changes made to the HAVING clause are rolled back by
THD::rollback_item_tree_changes(), but it doesn't restore the
used_tables() cache. The stale cache makes used_tables() return
incorrect values in the second execution.

The bug is fixed by making the hypergraph optimizer invoke
update_used_tables() on the HAVING clause before it analyzes it and
creates the corresponding FILTER access path.

Change-Id: I8dd266cb7ae3df74f09716b5f41a859e9d90010a