Bug #33768149 Clean up usage of init_sql_alloc
Remove the pre_alloc_size argument to init_sql_alloc, it has been
ignored since the patch for
    Bug #26940369: REWRITE THE MEM_ROOT

Remove calls to init_sql_alloc() for MEM_ROOT objects on the heap
(part of other objects) or on the stack. Use CTOR to initialize
instead.

MEM_ROOT now has a DTOR which does Clear(), so most of the calls to
this function can be removed.

Set MEM_ROOT.m_error_handler if MYSQL_SERVER is defined.

Change-Id: Iaeaaff7f2e9743a353c802829cd62ee9c73c7710