WL#15073 Instrument transaction batch memory
Encapsulate usage of Thd_ndb::m_batch_mem_root and add functions for
copying data into it.

Basically the transaction batching works by counting the estimated
number of bytes that will be sent to NDB when writing rows. Each time
another operation is added to the transaction, the unsent bytes counter
is increased. When the unsent bytes counter exceeeds the limit configured
by @@ndb_batch_size, the unsent parts of the transaction will be sent to
NDB and thus cause one roundtrip. Please note that the estimated number
of bytes does not exclude columns not written nor include the full data
size of blob columns, just the fixed size part of the blob. This patch
improves comments related to this functionality.

Change-Id: Iaac5db5c0d914fc2041f678a50ed92cee02d4838