Bug#31477725 NDB_RESTART FAILS SPORADICALLY IN PB2
Problem:
--------

The ndb_restart test verifies that the MySQL Server detects
system or node restarts (initial or otherwise) by grepping
the contents of the error logs. The test failed sporadically
with the grep in the second MySQL Server unable to find a
matching line as expected. This was due to a missing wait
for the second MySQL Server to reconnect.

Fix:
----

Add the missing wait for MySQL Server 2 to reconnect similar
to the one already present for MySQL Server 1.

Change-Id: I97d29f57c45834bd9adb50000588d0f886a8f0b3