Bug#33729162: Hypergraph misses rows for SELECT DISTINCT ... GROUP BY WITH ROLLUP
SELECT DISTINCT combined with GROUP BY WITH ROLLUP sometimes gave
wrong results with the hypergraph optimizer if the query was grouped
on the same expression multiple times through different aliases.

The problem was that it believed the aliases referenced equivalent
expressions, even though they were on different rollup levels, and
erroneously removed some of the expressions from the sort that did the
duplicate removal for DISTINCT.

The bug is fixed by making Item_rollup_group_item::eq() take the
min_rollup_level() into account when determining if two rollup group
items are equivalent.

Change-Id: Id47ac04f74f51620ac6c50f8bf6c5c9ec4f3bed7