BUG#27857780 PB2 PERFSCHEMA.START_SERVER_NO_MDL UNSTABLE AND FAILS SPORADICALLY
Problem:

Test perfschema.start_server_no_mdl is unstable.

Root cause:

The test attempts to verify that, after a FLUSH STATUS,
the counter PERFORMANCE_SCHEMA_METADATA_LOCK_LOST has a predictable value.

This is too optimistic, and assumes that the execution of:
- FLUSH STATUS
- the following SELECT FROM performance_schema.global_status
always takes the same number of metadata locks during execution,
and that nothing else in the server also takes metadata locks.

This is not stable, because of:
- possible usage of the audit log
- possible usage of the slow log
- possible usage of the general log
- various caches in the server
- possible execution of the event scheduler
- possible execution of other background threads
which all might use metadata locks,
causing the lost counter to increase in unpredictable ways.

Fix:

Abandon that part of the test,
testing FLUSH STATUS is only a nice to have,
but not practical.

Checking that the PERFORMANCE_SCHEMA_METADATA_LOCK_LOST indeed
increases when the mdl instrumentation is disabled is the
only critical functionality to test, and is covered.

Change-Id: I62434081a37ef1be8ffede90d41866ac57a42763