Bug#33753245 mysqld crashes in query with WINDOW function and WITH ROLLUP modifier
The rollup wrapper is missing for an aggregate induced into the select
list from the ORDER BY clause of a window tion.

Query_block::prepare detects when such aggregates are induced by the
HAVING clause, and also pessimistically assumes it may have been
induced always if there is an ORDER BY clause to the query, cf. the
variable `added_new_sum_funcs'.

The patch sets this to true if setting up windows increases the size
of fields by adding a hidden aggregate. This ensures we will look for
such aggregates and wrap the aggregate in a rollup wrapper if ROLLUP
is present.

Change-Id: I5c626150983292de7124a0a3a0738367ecea0148