Bug#33722819: recent regression: crash/corruption with gcols and multitable delete...
Some multi-table DELETE statements on tables with generated columns
crashed.

This problem was introduced when bug#33364732 made iterators for
DELETE operations. The specific change that caused trouble, was that
the code that added extra columns needed for delete to the read
set (TABLE::mark_columns_needed_for_delete()) was moved into
DeleteRowsIterator's constructor.

Previously, that modification to the read set happened before the
iterators were created. DynamicRangeIterator also fiddles with the
read set in its constructor. Because the iterator tree is constructed
bottom-up, DynamicRangeIterator's constructor runs before
DeleteRowsIterator's constructor. DynamicRangeIterator therefore
doesn't see the modifications DeleteRowsIterator has done to the read
set, and it skips the columns DeleteRowsIterator adds.

If DynamicRangeIterator::Init() later decides that it should use the
read set it created in the constructor, it ends up undoing the changes
DeleteRowsIterator made to the read set, and this causes crashes when
performing the deletes.

This patch moves the modification of the read set, and some other
related operations, out of DeleteRowsIterator's constructor and into a
separate function that is called prior to the creation of the iterator
tree, so that the read set again is extended with the extra columns
for DELETE before DynamicRangeIterator sees it.

Change-Id: I9a760732c1242c0e815680657950585bce8c9d94